grpc-lab/requirements.txt
Flask
grpcio
grpcio-tools
grpcio-reflection
protobuf
